[+] refactor : add resource overview component
